de-duplicate teleop scripts
