+++ agents/locobot/teleop.py
@@ -11,70 +11,19 @@ import functools
 from multiprocessing import set_start_method
 
 from droidlet import dashboard
+from droidlet.dashboard.o3dviz import o3dviz
 import numpy as np
 from scipy.spatial import distance
-
-os.environ["WEBRTC_IP"] = "0.0.0.0"
-os.environ["WEBRTC_PORT"] = "8889"
-
 import open3d as o3d
-o3d.visualization.webrtc_server.enable_webrtc()
 
-from open3d.visualization import O3DVisualizer, gui
-import threading
-import queue
 import time
 import math
-
-class O3dViz(threading.Thread):
-    def __init__(self, *args, **kwargs):
-        self.q = queue.Queue()
-        super().__init__(*args, **kwargs)
-
-    def put(self, name, command, obj):
-        # pass
-        self.q.put([name, command, obj])
-
-    def run(self):        
-        app = gui.Application.instance
-
-        app.initialize()
-        w = O3DVisualizer("o3dviz", 1024, 768)
-        w.set_background((0.0, 0.0, 0.0, 1.0), None)
-         
-        app.add_window(w)
-        reset_camera = False
-
-        while True:
-            app.run_one_tick()
-            time.sleep(0.001)
-
-            try:
-                name, command, geometry = self.q.get_nowait()
-
-                try:
-                    if command == 'remove':
-                        w.remove_geometry(name)
-                    elif command == 'replace':
-                        w.remove_geometry(name)
-                        w.add_geometry(name, geometry)
-                    elif command == 'add':
-                        w.add_geometry(name, geometry)
-
-                except:
-                    print("failed to add geometry to scene")
-                if not reset_camera:
-                    w.reset_camera_to_default()
-                    reset_camera = True
-                w.post_redraw()
-            except queue.Empty:
-                pass
-
 
 if __name__ == "__main__":
     # this line has to go before any imports that contain @sio.on functions
     # or else, those @sio.on calls become no-ops
     dashboard.start()
+    o3dviz.start()
 
 from droidlet.interpreter.robot import (
     dance, 
@@ -84,7 +33,6 @@ from droidlet.interpreter.robot import (
     LocoInterpreter,
 )
 from droidlet.dialog.robot import LocoBotCapabilities
-from droidlet.lowlevel.locobot.locobot_mover import LoCoBotMover
 from droidlet.event import sio
 
 faulthandler.register(signal.SIGUSR1)
@@ -99,42 +47,137 @@ logging.getLogger().handlers.clear()
 
 mover = None
 
-@sio.on("movement command")
-def test_command(sid, commands, yaw_velocity):
-    print(commands, yaw_velocity)
+@sio.on("sendCommandToAgent")
+def get_command(sid, command):
+    command, value = command.split()
+    print(command)
+    print(value)
+    test_command(sid, [command], value=value)
+
+@sio.on("logData")
+def log_data(sid, seconds):
+    test_command(sid, ["LOG_DATA"], value=seconds)
+
+@sio.on("stopRobot")
+def stop_robot(sid):
+    test_command(sid, ["STOP_ROBOT"])
+
+@sio.on("unstopRobot")
+def unstop_robot(sid):
+    test_command(sid, ["UNSTOP_ROBOT"])
+
+
+def test_command(sid, commands, data={"yaw": 0.1, "velocity": 0.1, "move": 0.3}, value=None):
+    print(commands, data, value)
+    move_dist = float(data['move'])
+    yaw = float(data['yaw'])
+    velocity = float(data['velocity'])
     global mover
     if mover == None:
         return
+    if value is not None:
+        move_dist = value
+
+    def sync():
+        time.sleep(10)
+        for i in range(50):
+            mover.get_rgb_depth()
+
     movement = [0.0, 0.0, 0.0]
     for command in commands:
         if command == "MOVE_FORWARD":
-            movement[0] += 0.1
-            print("action: FORWARD")
+            movement[0] += float(move_dist)
+            print("action: FORWARD", movement)
+            mover.move_relative([movement], blocking=False)
         elif command == "MOVE_BACKWARD":
-            movement[0] -= 0.1
-            print("action: BACKWARD")
+            movement[0] -= float(move_dist)
+            print("action: BACKWARD", movement)
+            mover.move_relative([movement], blocking=False)
         elif command == "MOVE_LEFT":
-            movement[2] += 0.3
-            print("action: LEFT")
+            movement[2] += yaw
+            print("action: LEFT", movement)
+            mover.move_relative([movement], blocking=False)
         elif command == "MOVE_RIGHT":
-            movement[2] -= 0.3
-            print("action: RIGHT")
+            movement[2] -= yaw
+            print("action: RIGHT", movement)
+            mover.move_relative([movement], blocking=False)
         elif command == "PAN_LEFT":
-            mover.bot.set_pan(mover.bot.get_pan() + 0.08)
+            mover.bot.set_pan(mover.get_pan() + yaw)
+            sync()
         elif command == "PAN_RIGHT":
-            mover.bot.set_pan(mover.bot.get_pan() - 0.08)
+            mover.bot.set_pan(mover.get_pan() - yaw)
+            sync()
         elif command == "TILT_UP":
-            mover.bot.set_tilt(mover.bot.get_tilt() - 0.08)
+            mover.bot.set_tilt(mover.get_tilt() + yaw)
+            print("action: TILT_UP", mover.get_tilt() + yaw)
+            sync()
         elif command == "TILT_DOWN":
-            mover.bot.set_tilt(mover.bot.get_tilt() + 0.08)
-        print(command, movement)
-        mover.move_relative([movement])
+            mover.bot.set_tilt(mover.get_tilt() - yaw)
+            sync()
+        elif command == "LOG_DATA":
+            mover.log_data_start(float(value)) # in seconds
+        elif command == "STOP_ROBOT":
+            mover.stop()
+        elif command == "UNSTOP_ROBOT":
+            mover.unstop()
+        elif command == "SET_PAN":
+            print("action: SET_PAN", float(value))
+            mover.bot.set_pan(float(value))
+            sync()
+        elif command == "SET_TILT":
+            print("action: SET_TILT", float(value))
+            mover.bot.set_tilt(float(value))
+            sync()
+        elif command == "MOVE_ABSOLUTE":
+            xyyaw_s = value.split(',')
+            xyyaw_f = [float(v) for v in xyyaw_s]
+            print("action: MOVE_ABSOLUTE", xyyaw_f)
+            mover.move_absolute(xyyaw_f, blocking=False)
+            sync()
+        elif command == "LOOK_AT":
+            xyz = value.split(',')
+            xyz = [float(p) for p in xyz]
+            print("action: LOOK_AT", xyz)
+            mover.look_at(xyz, turn_base=False)
+        elif command == "RESET":
+            mover.bot.set_tilt(0.)
+            mover.bot.set_pan(0.)
 
+        print(command, movement)
+
+@sio.on("movement command")
+def test_command_web(sid, commands, data, value=None):
+    test_command(sid, commands, data=data, value=value)
 
 if __name__ == "__main__":
-    ip = os.getenv("LOCOBOT_IP")
+    import argparse
+
+    parser = argparse.ArgumentParser(description="Pass in server device IP")
+    parser.add_argument(
+        "--ip",
+        help="Server device (robot) IP. Default is 0.0.0.0",
+        type=str,
+        default="0.0.0.0",
+    )
+    parser.add_argument(
+        "--backend",
+        help="Which backend to use: habitat (default), hellorobot",
+        type=str,
+        default='habitat',
+    )
+    args = parser.parse_args()
+    
+    ip = args.ip
+    backend = args.backend
+    
     print("Connecting to robot at ip: ", ip)
-    mover = LoCoBotMover(ip=ip)
+
+    if backend == 'habitat':
+        from droidlet.lowlevel.locobot.locobot_mover import LoCoBotMover
+        mover = LoCoBotMover(ip=ip, backend='habitat')
+    elif backend == 'hellorobot':
+        from droidlet.lowlevel.hello_robot.hello_robot_mover import HelloRobotMover
+        mover = HelloRobotMover(ip=ip)
     print("Mover is ready to be operated")
 
     log_settings = {
@@ -142,37 +185,34 @@ if __name__ == "__main__":
         "image_quality": 10,  # from 10 to 100, 100 being best
     }
 
-    o3dviz = O3dViz()
-    o3dviz.start()
-
     all_points = None
     all_colors = None
     first = True
     prev_stg = None
     path_count = 0
 
-    start_time = time.time()
+    start_time = time.time_ns()
     fps_freq = 1 # displays the frame rate every 1 second
     counter = 0
     
     while True:
         counter += 1
-        if (time.time() - start_time) > fps_freq :
-            print("FPS: ", counter / (time.time() - start_time))
+        iter_time = time.time_ns() - start_time
+        if float(iter_time) / 1e9 > fps_freq :
+            print("FPS: ", round(counter / (float(iter_time) / 1e9), 1), "  ", int(iter_time / 1e6 / counter), "ms")
             counter = 0
-            start_time = time.time()
-
-        movement = [0.0, 0.0, 0.3]
-        mover.move_relative([movement])
+            start_time = time.time_ns()
 
         base_state = mover.get_base_pos_in_canonical_coords()
-        print("base_state: ", base_state)
-        print("rgb_depth: ", mover.get_rgb_depth())
+
         sio.emit("image_settings", log_settings)
         resolution = log_settings["image_resolution"]
         quality = log_settings["image_quality"]
 
+        # this goes from 21ms to 120ms
         rgb_depth = mover.get_rgb_depth()
+
+        # this takes about 1.5 to 2 fps
         serialized_image = rgb_depth.to_struct(resolution, quality)
 
         sio.emit("rgb", serialized_image["rgb"])
@@ -182,9 +222,9 @@ if __name__ == "__main__":
             "depthMin": serialized_image["depth_min"],
         })
 
-        pcd = mover.get_current_pcd(in_global=True)
-        points = pcd[0]
-        colors = pcd[1]/255.
+
+        points, colors = rgb_depth.ptcloud.reshape(-1, 3), rgb_depth.rgb.reshape(-1, 3)
+        colors = colors / 255.
 
         if all_points is None:
             all_points = points
@@ -198,21 +238,21 @@ if __name__ == "__main__":
         opcd.colors = o3d.utility.Vector3dVector(all_colors)
         opcd = opcd.voxel_down_sample(0.05)
 
-
-        # remove the rooftop / ceiling points in the point-cloud to make it easier to see the robot in the visualization
-        crop_bounds = o3d.utility.Vector3dVector([
-            [-1000., -1000., -1000.],
-            [1000., 1000., 2.0],
-            ])
-        opcd = opcd.crop(
-            o3d.geometry.AxisAlignedBoundingBox.create_from_points(
-                crop_bounds,
-            )
-        )
-
+        # # remove the rooftop / ceiling points in the point-cloud to make it easier to see the robot in the visualization
+        # crop_bounds = o3d.utility.Vector3dVector([
+        #     [-1000., -20., -1000.],
+        #     [1000., 20., 1000.0],
+        #     ])
+        # opcd = opcd.crop(
+        #     o3d.geometry.AxisAlignedBoundingBox.create_from_points(
+        #         crop_bounds,
+        #     )
+        # )
+        
+        
         all_points = np.asarray(opcd.points)
         all_colors = np.asarray(opcd.colors)
-
+        
         if first:
             cmd = 'add'
             first = False
@@ -220,7 +260,6 @@ if __name__ == "__main__":
             cmd = 'replace'
             
         o3dviz.put('pointcloud', cmd, opcd)
-
 
         # Plot the robot
         x, y, yaw = base_state.tolist()
@@ -231,78 +270,85 @@ if __name__ == "__main__":
                                                        cone_height = .4,
                                                        resolution=20)
         robot_orientation.compute_vertex_normals()
-        robot_orientation.paint_uniform_color([0.1, 0.9, 0.1])
+        robot_orientation.paint_uniform_color([1.0, 0.5, 0.1])
         
-        robot_orientation.translate([y, -x, 0.5], relative=False)
+        robot_orientation.translate([y, -x, 0.], relative=False)
+        # make the cylinder representing the robot to be parallel to the floor
         robot_orientation.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, math.pi/2, 0]))
-        robot_orientation.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, 0, yaw]))        
+        # rotate the cylinder by the robot orientation
+        if yaw != 0:
+            robot_orientation.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, 0, yaw]))
 
         o3dviz.put('bot_orientation', cmd, robot_orientation)
 
         robot_base = o3d.geometry.TriangleMesh.create_cylinder(radius=.1,
-                                                          height=1.,)
-        robot_base.translate([y, -x, 0.4], relative=False)
+                                                          height=1,)
+        robot_base.translate([y, -x, 0.1], relative=False)
         robot_base.compute_vertex_normals()
         robot_base.paint_uniform_color([1.0, 1.0, 0.1])
 
         o3dviz.put('bot_base', cmd, robot_base)
 
+        # red = x, green = y, blue = z
+        axis = o3d.geometry.TriangleMesh.create_coordinate_frame(size=1.0, origin=np.array([0., 0., 0.]))        
+        axis.compute_vertex_normals()
+        o3dviz.put('axis', cmd, axis)
 
         # start the SLAM
-        mover.explore()
+        if backend == 'habitat':
+            mover.explore()
         
-        # get the SLAM goals
-        goal_loc, stg = None, None # mover.bot.get_slam_goal()    
+            # get the SLAM goals
+            goal_loc, stg = None, None # mover.bot.get_slam_goal()    
 
-        # plot the final goal
-        if goal_loc is not None:
-            goal_x, goal_y, goal_z = goal_loc
-            cone = o3d.geometry.TriangleMesh.create_cylinder(radius=.2,
-                                                             height=3.,)
-            cone.translate([goal_x, goal_y, 0.4], relative=False)
-            cone.compute_vertex_normals()
-            cone.paint_uniform_color([0.0, 1.0, 1.0])
-            o3dviz.put('goal_cone', cmd, cone)
+            # plot the final goal
+            if goal_loc is not None:
+                goal_x, goal_y, goal_z = goal_loc
+                cone = o3d.geometry.TriangleMesh.create_cylinder(radius=.2,
+                                                                 height=3.,)
+                cone.translate([goal_x, goal_y, 0.4], relative=False)
+                cone.compute_vertex_normals()
+                cone.paint_uniform_color([0.0, 1.0, 1.0])
+                o3dviz.put('goal_cone', cmd, cone)
 
-        # plot the short term goal in yellow and the path in green
-        if stg is not None:
-            stg_x, stg_y = stg
-            cone = o3d.geometry.TriangleMesh.create_cylinder(radius=.2,
-                                                             height=3.,)
-            cone.translate([stg_x, stg_y, 1.4], relative=False)
-            cone.compute_vertex_normals()
-            cone.paint_uniform_color([1.0, 1.0, 0.0])
-            o3dviz.put('stg', cmd, cone)
+            # plot the short term goal in yellow and the path in green
+            if stg is not None:
+                stg_x, stg_y = stg
+                cone = o3d.geometry.TriangleMesh.create_cylinder(radius=.2,
+                                                                 height=3.,)
+                cone.translate([stg_x, stg_y, 1.4], relative=False)
+                cone.compute_vertex_normals()
+                cone.paint_uniform_color([1.0, 1.0, 0.0])
+                o3dviz.put('stg', cmd, cone)
 
-            if prev_stg is None:
-                prev_stg = [y, -x]
-            cur_stg = [stg_x, stg_y]
+                if prev_stg is None:
+                    prev_stg = [y, -x]
+                cur_stg = [stg_x, stg_y]
 
-            arrow_length = distance.euclidean(cur_stg, prev_stg)
-            if arrow_length > 0.0001:                
-                path = o3d.geometry.TriangleMesh.create_arrow(cylinder_radius=.03,
-                                                              cone_radius=.04,
-                                                              cylinder_height = arrow_length / 2,
-                                                              cone_height = arrow_length / 2,)
-                path.compute_vertex_normals()
-                path.paint_uniform_color([0.0, 1.0, 0.0])
-        
-                path.translate([prev_stg[0], prev_stg[1], 0.2], relative=False)
-                path.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, math.pi/2, 0]))
-                path.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, 0, yaw]))        
-                o3dviz.put('short_term_goal_path_{}'.format(path_count), 'add', path)
-                path_count = path_count + 1
-            prev_stg = cur_stg
+                arrow_length = distance.euclidean(cur_stg, prev_stg)
+                if arrow_length > 0.0001:                
+                    path = o3d.geometry.TriangleMesh.create_arrow(cylinder_radius=.03,
+                                                                  cone_radius=.04,
+                                                                  cylinder_height = arrow_length / 2,
+                                                                  cone_height = arrow_length / 2,)
+                    path.compute_vertex_normals()
+                    path.paint_uniform_color([0.0, 1.0, 0.0])
 
-        # # get the obstacle map and plot it
-        # obstacles = mover.bot.get_map()
-        # obstacles = np.asarray(obstacles)
-        # obstacles = np.concatenate((-obstacles[:, [1]], -obstacles[:, [0]], np.zeros((obstacles.shape[0], 1))), axis=1)
-        # obspcd = o3d.geometry.PointCloud()
-        # obspcd.points = o3d.utility.Vector3dVector(obstacles)
-        # obspcd.paint_uniform_color([1.0, 0., 0.])
-        # obsvox = o3d.geometry.VoxelGrid.create_from_point_cloud(obspcd, 0.03)
-        # o3dviz.put('obstacles', cmd, obsvox)                
-            
+                    path.translate([prev_stg[0], prev_stg[1], 0.2], relative=False)
+                    path.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, math.pi/2, 0]))
+                    path.rotate(o3d.geometry.get_rotation_matrix_from_axis_angle([0, 0, yaw]))        
+                    o3dviz.put('short_term_goal_path_{}'.format(path_count), 'add', path)
+                    path_count = path_count + 1
+                prev_stg = cur_stg
+
+            # # get the obstacle map and plot it
+            # obstacles = mover.bot.get_map()
+            # obstacles = np.asarray(obstacles)
+            # obstacles = np.concatenate((-obstacles[:, [1]], -obstacles[:, [0]], np.zeros((obstacles.shape[0], 1))), axis=1)
+            # obspcd = o3d.geometry.PointCloud()
+            # obspcd.points = o3d.utility.Vector3dVector(obstacles)
+            # obspcd.paint_uniform_color([1.0, 0., 0.])
+            # obsvox = o3d.geometry.VoxelGrid.create_from_point_cloud(obspcd, 0.03)
+            # o3dviz.put('obstacles', cmd, obsvox)
         
         time.sleep(0.001)
